+++ frame/message-queue/src/mock_helpers.rs
@@ -112,7 +112,7 @@ pub fn book_for<T: Config>(page: &PageOf<T>) -> BookStateOf<T> {
 		end: 1,
 		ready_neighbours: None,
 		message_count: page.remaining.into(),
-		size: page.remaining_size.into(),
+		size: page.remaining_size.into() as u64,
 	}
 }
